Fixed path search for data folder
As the directory for the config file is now installed with the
modules, it was necessary to use the appropriate path relative
identification to extract and save to the right location.
This was acheived with the pkg_resources module. Previous
attempts are still in the script comments, in case future bugs
arise that can be solved through them.