added best match functionality and error handling to geneid parsing
for the My_gff methods whichstrand and whichscaf, the passing of
a non-existent gene previously led to a KeyError. Now it will
return None if it fails to match a gene. In addition, setting the
'exact' parameter to False will allow the class to search for all
genes that match the input, and if a single match (and only a
single match) is found, it will return the result for that match.