improved efficiency in genematch.extractseq function
removed two of the three regex searches that occurred each line
of the target file (relying on the '>' marker of fasta format
files being in the first position of a defline).

Also modified the regex search for the gene name to be more
sensitive to variable deflines. Now any line that contains the
target string followed by \w or \n will match, so it does not
rely on a strict format in the defline for extracting the target
feature.