reduced dependencies, added options and functionality
Removed the code using the statusbar module and wrote in my own,
to reduce the number of dependencies required to run gffparser.
TODO: will add my own statusbar script into config to allow access
by other modules.

Added the option to run the GO analyses during methylation
analyses. It was previously mandatory, but is not always useful
and it takes a lot longer if turned on.

Added the config logging functionality

clarified the variable used in My_gff.ingene() is gene ID and not
gene name

added verbalise functionality to some of the output

cleaned the argparser section for easier human reading

modified snp_in_gene(), so that it recognises pseudo- and partial-
genes, and subsequently ignores them (but at least does not error)
Also, this function now catches the error that occurs if the gene
is bigger than the mRNA location, such that a SNP is registered as
being "in" the gene, but then not in any of the sub-features.
Because these instances are currently caught by the snp_in_gene
feature downstream, they could be safely ignored here, but a more
robust re-coding should be done to return the internal (intronic,
5' UTR, etc) position of these SNPs.