improved My_gff class with added feature analysis
tidied up the code and creation of the __init__ method of the
class, allowing more user choice in which level a given instance
will parse out of the gff file (eg. gene or mRNA level). Also
added dictionaries to store the multi-level features of the NCBI
format with multiple isoforms per gene (even if mRNA is the upper
level chosen for an instance, it will still store information
about which gene each mRNA comes from).

Also added a gff line parser function to quickly gain access to
all the attributes. Not sure why this function has taken so long
to be added to a gff parsing module!!