added file overwrite checking and option
when an output directory and filename were explicitly given in
the past, config logging would not check to see if the file
already existed, assuming that the time-stamp that was added
would lead to a novel filename. However, if the program is run
again within a minute, the time stamp is the same, so the file
name would be duplicated and the first file overwritten.

When default settings were used, and the program was similarly
run within a minute of the last time, it would recognise that
the file already existed, but would then error and stop.

So to get around both these bugs, the config module now always
checks to see if the new file exists, and also gives the option
(if it does) to overwrite the existing file.