added creation of mastergene to gff class
For analyses that require a single gene model to identify
equivalent exons, or look at the total size of all contributing
exons, it is necessary to create a 'master gene', consisting of
a merge of all isoforms. This was in this instance accomplished by
extracting all exons for all isoforms of each gene, then pooling
all overlapping exons and taking the extreme bounds of the pooled
exons, to create the maximal extent of each possible exon for a
gene.