changed FDR calculation! Also added ability to randomly iterate.
The change that will affect most programs is the changing of the
p_to_q function from the Storey/Tibshirani FDR method to now
accessing statsmodel FDR methods. The default is now Benjamini/
Hochbergs FDR, although multiple other methods are now accessible.
The old ST FDR method is still available as p_to_q_storey_tibshirani
but since this method struggles with the P value distribution
arisign from Fisher's Exaact Tests (due to a large proportion of
P = 1.0), the otherwise more conservative Benjamini Hochberg method
actually provides a more robust analysis, particularly for smaller
distributions.

Genematch also now has the ability to perform enrichment on a
randomly selected subset of the input gene list (subset size is
specified by the user). In addition, the enrichment analysis can
be repeated by a user-specified number of iterations. This was
incorporated to allow comparison of enrichment results with that
of a random subset from the genome.

Although primarily to benefit the use of the StoreyTibshirani FDR
function, an option has also been added to allow the user to
specify whether or not to display graphs (so that it can be set
independently of the quiet console output).