added plotting of results, updated to run with GSNAP v2015-06-10
new function and menu option allows user to plot the indel rate per
base position from the stats files. This can be run in the complete
pipeline (after building the reference, aligning the reads, and
generating the stats), or you can point the directory to the location
of the stats files and disable alignment and just visualise the
stats files already prepared.

Also updated the align_reads function to be compatible with
GSNAP version 2015-06-10: A memory error continued to pop up
relating to shmget, so to avoid the error the use of shared memory
has been turned off. This should still be backwards compatible,
though may run a little slower than in the past.