smarter checking of path option for make_a_list()
rather than assuming any string containing backslashes is refering
to a file that needs to be opened, make_a_list now uses the os
function to check whether a string points to a valid path. Otherwise,
it will treat it as a string.

also updated the cfg dictionary keys to reflect latest additions to
the hymenoptera ortholog database.