Major additions to GffLibrary class and creation of FastaLib class
FastaLib class allows association of genome assembly nucleotide
sequences to the GffLibrary class, for extraction of sequences
associated with gff features.

All gff features now receive a unique id : the CDS features under the
NCBI GFF3 format do not have unique IDs assigned, so the class now
checks if the feature type is CDS, and then appends the start field
to the id, creating an id unique for that CDS exon.

NEW extractseq() method:

this method accesses the FastaLib class to obtain the relevant
nucleotide sequence associated with a gff feature. Parameters
currently available include:
   trim_from and trim_to : allowing partial sequence collection
       (which will be used for creating PCR products)
   boundaries boolean    : allows return of an ordered list of
       exon lengths (also to be used for qPCR design)
   cds or genomic        : can specify whether sequence returned is
       entire genomic sequence for region, or just cds
   buffer                : for genomic sequence, can extract
       sequence on either side of the feature
   translation           : can return the peptide translation of
       the extracted nucleotide sequence

NEW _get_cdfeatures() method

creates a list of isoform lists, each containing the CDS features.

NEW _get_seq() method

used by the above methods for the actual sequence extraction (slicing
occurs afterwards in the necessary method)

NEW fetch_promoter() method

Similar to extractseq, but more limited in output. This method will
return sequences a specified distance upstream and specified distance
downstream of the start codon. This method is also accessible via
the command line arguments, to output a list of promoter sequences
for a specified list of genes.

NEW find_nearest() method

This is functionally equivalent to the same method in the old gff
class, allowing the identification of the closes features up- and
down- stream of the specified locus.