allow use of config file keys in specifying path of reference file.
To better utilise the fact that the config file already contains
most of the paths needed to locate genome reference files, the args
for specifying the reference path can now also reference the key
variable in the config file, which will be converted to the full
path.

Also moved the initialising of dbpaths to __main__, and removed all
dbpath references from function default variables.

Also moved argument parser variables from __main__ to a function,
to be consistent with the other genomepy modules.