updated GffLIbrary object to return CDS sequence only
when a gene name is given to Cris, the GffLibrary is used to
extract the gene sequence information. This had recently been
updated from the old GFF parser class, but the extractseq method
did not have the cds parameter set to True. This has now been
corrected, leading to more efficient (and much, much shorter)
searches.